requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
mysqlclient==2.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
WTForms==3.0.1
email-validator==2.0.0
cryptography==41.0.4
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2 